refactor(layout, NgRx): move `initial view selector` to `layout`
