+++ src/app/core/layout/layout.component.ts
@@ -6,7 +6,7 @@ import { Observable } from 'rxjs';
 import { FooterAction } from '../models/footer-actions.model';
 import { IInitialViewModel } from '../models/initial-view-model.model';
 import { coreActions } from '@fts-store/core';
-import { selectInitialViewModel } from '@fts-store/core';
+import { selectInitialViewModel } from '@fts-store/layout';
 import { FooterActionsComponent } from './components/footer-actions/footer-actions.component';
 import { MainMenuComponent } from './components/main-menu/main-menu.component';
 import { PageHeaderComponent } from './components/page-header/page-header.component';
