+++ src/app/core/store/core/index.ts
@@ -1,3 +1,2 @@
 export * from './core-message.effects';
 export * from './core.actions';
-export * from './initial.selectors';
