+++ src/app/core/store/layout/index.ts
@@ -1,3 +1,4 @@
 export * from './layout.feature';
 export * from './layout.actions';
 export * from './layout.effects';
+export * from './layout.selectors';
